+++ src/app/header/header.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import {ProductListComponent} from "../product/product-list/product-list.component";
+import {HeaderComponent} from "./header.component";
+
+@NgModule({
+  imports: [
+    CommonModule
+  ],
+  declarations: [HeaderComponent],
+  exports: [HeaderComponent]
+})
+export class HeaderModule { }