+++ src/app/product/product.module.ts
@@ -1,13 +1,15 @@
 import {NgModule} from '@angular/core';
+import {FormsModule} from "@angular/forms";
 import {CommonModule} from '@angular/common';
 
 import {ProductItemComponent} from './product-item/product-item.component';
 import {ProductService} from './product.service';
 import {ProductListComponent} from "./product-list/product-list.component";
+import {HeaderModule} from "../header/header.module";
 
 @NgModule({
   imports: [
-    CommonModule
+    CommonModule, FormsModule, HeaderModule
   ],
   declarations: [ProductListComponent, ProductItemComponent],
   providers: [ProductService],